Fix export indexing bug
The top unique ranking algorithm had to be updated. Instead of having
the top 'n' be the individual words, we check to make sure that there
are at least 'n' ranks, and then return a list with full indices (no
empty spaces due to not having enough ranks to return). Now the final
export list doesn't have empty spots that crash the program.