Fix the unique finding algorithm.
The old version of the unique audio/video finding algorithm had a
workaround for an old bug that was causing a non-unique word to make it
to the final results. This loophole only allowed the word with the
highest count to make it in (regardless of whether or not it was
unique). Closed the loophole.